Update patch set 4
Patch Set 4:

The Google Webview doesn't contain an xxxhdpi drawable? Was that svg linked earlier actually a vector image or just an svg with an embedded raster image? If the former you can use that to create an xxxhdpi drawable.

I didn't check the drawables since I only have access to my phone. I won't be able to do so for a day or two probably.

Patch-set: 4