Update patch set 5
Patch Set 5:

Joey: Unless you redraw it as a vector or use a vectorizer and fix all the mistakes by hand, putting it in inkscape is pointless.

No, we don't really need xxxhdpi for this right now since it won't be shown in the launcher. Otherwise we would. Remove the upscaled drawable though.

Patch-set: 5